@@ -6,7 +6,7 @@ homepage: https://github.com/hyiso/lint_staged
 repository: https://github.com/hyiso/lint_staged
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
   ansi: ^0.2.2
@@ -18,7 +18,7 @@ dependencies:
   yaml: ^3.1.1
 
 dev_dependencies:
-  commitlint_cli: ^0.6.0-pre.0
+  commitlint_cli: ^0.6.1
   husky: ^0.1.6
   lints: ^2.0.0
   test: ^1.16.0
