@@ -1,18 +1,30 @@
 import 'dart:io';
 
 import 'package:path/path.dart';
+import 'package:verbose/verbose.dart';
 import 'package:yaml/yaml.dart';
 
-Future<Map<String, List<String>>?> loadConifg({
+final _verbose = Verbose('lint_staged:config');
+
+Future<Map<String, List<String>>?> loadConfig({
   String? workingDirectory,
 }) async {
   final pubspecPath =
       join(workingDirectory ?? Directory.current.path, 'pubspec.yaml');
   final yaml = await loadYaml(File(pubspecPath).readAsStringSync());
-  if (yaml['lint_staged'] is! Map) {
+  final map = yaml['lint_staged'];
+  if (map is! Map) {
     return null;
   }
-  final config = yaml['lint_staged'] as Map;
-  return config.cast<String, String>().map<String, List<String>>((key, value) =>
-      MapEntry(key, value.split('&&').map((e) => e.trim()).toList()));
+  _verbose('Found config: $map');
+  final config = <String, List<String>>{};
+  for (var entry in map.entries) {
+    final value = entry.value;
+    if (value is String) {
+      config[entry.key] = value.split('&&').map((e) => e.trim()).toList();
+    } else if (value is List) {
+      config[entry.key] = value.cast();
+    }
+  }
+  return config;
 }
