@@ -1,153 +1,145 @@
 import 'dart:io';
 
-import 'package:glob/glob.dart';
-import 'package:lint_staged/src/logger.dart';
+import 'package:ansi/ansi.dart';
+import 'package:lint_staged/src/fs.dart';
 import 'package:verbose/verbose.dart';
 
 import 'chunk.dart';
 import 'config.dart';
-import 'exception.dart';
 import 'git.dart';
-import 'git_workflow.dart';
-import 'lint_runner.dart';
+import 'group.dart';
+import 'workflow.dart';
+import 'logging.dart';
 import 'message.dart';
 import 'context.dart';
 import 'symbols.dart';
 
-final verbose = Verbose('lint_staged:run');
+final _verbose = Verbose('lint_staged:run');
 
-Future<LintStagedContext> runAll({
+Future<Context> runAll({
   bool allowEmpty = false,
   List<String> diff = const [],
   String? diffFilter,
   bool stash = true,
   String? workingDirectory,
   int maxArgLength = 0,
-  required Logger logger,
 }) async {
   final ctx = getInitialContext();
-  if (!FileSystemEntity.isDirectorySync('.git') &&
-      !FileSystemEntity.isFileSync('.git')) {
+  final fs = FileSystem(workingDirectory);
+  if (!await fs.exists('.git')) {
     ctx.output.add(kNotGitRepoMsg);
     ctx.errors.add(kGitRepoError);
-    throw createError(ctx, kNotGitRepoMsg);
+    throw ctx;
   }
+  final git = Git(
+      diff: diff, diffFilter: diffFilter, workingDirectory: workingDirectory);
 
   /// Test whether we have any commits or not.
   /// Stashing must be disabled with no initial commit.
-  final hasInitialCommit =
-      await execGit(['log', '-1'], workingDirectory: workingDirectory)
-          .then((s) => true)
-          .catchError((s) => false);
+  final hasInitialCommit = (await git.lastCommit).isNotEmpty;
 
   /// lint_staged will create a backup stash only when there's an initial commit,
   /// and when using the default list of staged files by default
   ctx.shouldBackup = hasInitialCommit && stash;
   if (!ctx.shouldBackup) {
-    verbose(skippingBackupMsg(hasInitialCommit, diff));
+    final reason = diff.isNotEmpty
+        ? '`--diff` was used'
+        : hasInitialCommit
+            ? '`--no-stash` was used'
+            : 'there\'s no initial commit yet';
+    stdout.warn('Skipping backup because $reason.');
   }
-  final stagedFiles = await getStagedFiles(
-    diff: diff,
-    diffFilter: diffFilter,
-    workingDirectory: workingDirectory,
-  );
-  if (stagedFiles == null) {
-    ctx.output.add(kGetStagedFilesErrorMsg);
-    ctx.errors.add(kGetStagedFilesError);
-    throw createError(ctx, kNoStagedFilesMsg);
-  }
+  final stagedFiles = await git.stagedFiles;
   if (stagedFiles.isEmpty) {
     ctx.output.add(kNoStagedFilesMsg);
     return ctx;
   }
-  verbose('Loaded list of staged files int git:\n$stagedFiles');
-
-  final foundConfigs = await loadConifg(workingDirectory: workingDirectory);
-  if (foundConfigs == null) {
+  final config = await loadConfig(workingDirectory: workingDirectory);
+  if (config == null || config.isEmpty) {
     ctx.errors.add(kConfigNotFoundError);
-    throw createError(ctx, kNoConfigurationMsg);
+    throw ctx;
   }
-  if (foundConfigs.isEmpty) {
-    ctx.errors.add(kConfigEmptyError);
+  final groups = groupFilesByConfig(config: config, files: stagedFiles);
+  if (groups.isEmpty) {
+    ctx.output.add(kNoStagedFilesMatchedMsg);
     return ctx;
   }
-  verbose('Found configs: $foundConfigs');
-  final matchedFiles = <String>{};
-  final tasks = <Future Function()>[];
-  for (var pattern in foundConfigs.keys) {
-    final glob = Glob(pattern);
-    final scopedFiles = <String>{};
-    for (var file in stagedFiles) {
-      verbose('$file matches $pattern is ${glob.matches(file)}');
-      if (glob.matches(file)) {
-        scopedFiles.add(file);
-        matchedFiles.add(file);
-      }
-    }
-    stagedFiles.removeWhere((file) => scopedFiles.contains(file));
-    final runner = LintRunner(
-        ctx: ctx,
-        pattern: pattern,
-        fileList: scopedFiles.toList(),
-        scripts: foundConfigs[pattern]!,
-        workingDirectory: workingDirectory);
-    tasks.add(runner.run);
-  }
-  verbose('matched files: $matchedFiles');
+  final matchedFiles =
+      groups.values.expand((element) => element.files).toList();
   final matchedFileChunks =
-      chunkFiles(matchedFiles.toList(), maxArgLength: maxArgLength);
-  verbose('matched file chunks: $matchedFileChunks');
-  final git = GitWorkflow(
+      chunkFiles(matchedFiles, maxArgLength: maxArgLength);
+  final workflow = Workflow(
+    fs: fs,
+    ctx: ctx,
+    git: git,
     allowEmpty: allowEmpty,
-    gitConfigDir: await getGitConfigDir(),
-    diff: diff,
-    diffFilter: diffFilter,
     matchedFileChunks: matchedFileChunks,
-    workingDirectory: workingDirectory,
   );
-  SpinnerProgress progress = logger.progress('Preparing lint_staged...');
-  await git.prepare(ctx);
-  progress.finish();
+  final spinner = Spinner();
+  spinner.progress('Preparing lint_staged...');
+  await workflow.prepare();
+  spinner.success('Prepared lint_staged');
   if (ctx.hasPartiallyStagedFiles) {
-    progress =
-        logger.progress('Hiding unstaged changes to partially staged files...');
-    await git.hideUnstagedChanges(ctx);
-    progress.finish();
+    spinner.progress('Hide unstaged changes...');
+    await workflow.hideUnstagedChanges();
+    spinner.success('Hide unstaged changes');
+  } else {
+    spinner.skipped('Hide unstaged changes');
   }
-  if (matchedFiles.isNotEmpty) {
-    progress = logger.progress('Running tasks for staged files...');
-    await Future.wait(tasks.map((task) => task()));
-    progress.finish();
-    // logger.success('Running tasks for staged files...');
-  }
+  spinner.progress('Run tasks for staged files...');
+  await Future.wait(groups.values.map((group) async {
+    await Future.wait(group.scripts.map((script) async {
+      final args = script.split(' ');
+      final exe = args.removeAt(0);
+      await Future.wait(group.files.map((file) async {
+        final result = await Process.run(exe, [...args, file],
+            workingDirectory: workingDirectory);
+        final messsages = ['$script $file'];
+        if (result.stderr.toString().trim().isNotEmpty) {
+          messsages.add(ansi.red(result.stderr.toString().trim()));
+        }
+        if (result.stdout.toString().trim().isNotEmpty) {
+          messsages.add(result.stdout.toString().trim());
+        }
+        _verbose(messsages.join('\n'));
+        if (result.exitCode != 0) {
+          ctx.output.add(messsages.join('\n'));
+          ctx.errors.add(kTaskError);
+        }
+      }));
+    }));
+  }));
+  spinner.success('Run tasks for staged files');
   if (!applyModifationsSkipped(ctx)) {
-    progress = logger.progress('Applying modifications from tasks...');
-    await git.applyModifications(ctx);
-    progress.finish();
-    // logger.success('Applying modifications from tasks...');
+    spinner.progress('Apply modifications...');
+    await workflow.applyModifications();
+    spinner.success('Apply modifications');
+  } else {
+    spinner.skipped('Apply modifications');
   }
   if (ctx.hasPartiallyStagedFiles && !restoreUnstagedChangesSkipped(ctx)) {
-    progress = logger
-        .progress('Restoring unstaged changes to partially staged files...');
-    await git.resotreUnstagedChanges(ctx);
-    progress.finish();
-    // logger.success('Restoring unstaged changes to partially staged files...');
+    spinner.progress('Restore unstaged changes...');
+    await workflow.resotreUnstagedChanges();
+    spinner.success('Restore unstaged changes');
+  } else {
+    spinner.skipped('Restore unstaged changes');
   }
   if (restoreOriginalStateEnabled(ctx) && !restoreOriginalStateSkipped(ctx)) {
-    progress =
-        logger.progress('Reverting to original state because of errors...');
-    await git.restoreOriginState(ctx);
-    progress.finish();
-    // logger.success('Reverting to original state because of errors...');
+    spinner.progress('Revert because of errors...');
+    await workflow.restoreOriginState();
+    spinner.success('Revert because of errors');
+  } else {
+    spinner.skipped('Revert because of errors');
   }
   if (cleanupEnabled(ctx) && !cleanupSkipped(ctx)) {
-    progress = logger.progress('Cleaning up temporary files...');
-    await git.cleanup(ctx);
-    progress.finish();
-    // logger.success('Cleaning up temporary files...');
+    spinner.progress('Cleanup temporary files...');
+    await workflow.cleanup();
+    spinner.success('Cleanup temporary files');
+  } else {
+    spinner.skipped('Cleanup temporary files');
   }
   if (ctx.errors.isNotEmpty) {
-    throw createError(ctx);
+    throw ctx;
   }
   return ctx;
 }
