@@ -10,53 +10,52 @@ void main() {
         'fails when task reverts staged changes without `--allow-empty`, to prevent an empty git commit',
         () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatFix);
+      await project.fs.write('pubspec.yaml', kConfigFormatFix);
 
-      // Create and commit a pretty file without running lint_staged
+      // Create and commit a formatted file without running lint_staged
       // This way the file will be available for the next step
-      await project.writeFile('lib/main.dart', kFormattedDart);
-      await project.execGit(['add', '.']);
-      await project.execGit(['commit', '-m committed pretty file']);
+      await project.fs.write('lib/main.dart', kFormattedDart);
+      await project.git.run(['add', '.']);
+      await project.git.run(['commit', '-m committed formatted file']);
 
       // Edit file to be ugly
-      await project.removeFile('lib/main.dart');
-      await project.writeFile('lib/main.dart', kUnFormattedDart);
-      await project.execGit(['add', 'lib/main.dart']);
+      await project.fs.remove('lib/main.dart');
+      await project.fs.write('lib/main.dart', kUnFormattedDart);
+      await project.git.run(['add', 'lib/main.dart']);
 
       // Run lint_staged to automatically format the file
-      // Since prettier reverts all changes, the commit should fail
+      // Since formatter reverts all changes, the commit should fail
       await expectLater(project.gitCommit(), throwsException);
 
       // Something was wrong so the repo is returned to original state
+      expect(await project.git.commitCount, equals(2));
       expect(
-          await project.execGit(['rev-list', '--count', 'HEAD']), equals('2'));
-      expect(await project.execGit(['log', '-1', '--pretty=%B']),
-          contains('committed pretty file'));
-      expect(await project.readFile('lib/main.dart'), equals(kUnFormattedDart));
+          await project.git.lastCommit, contains('committed formatted file'));
+      expect(await project.fs.read('lib/main.dart'), equals(kUnFormattedDart));
     });
 
     test(
         'creates commit when task reverts staged changed and --allow-empty is used',
         () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatFix);
+      await project.fs.write('pubspec.yaml', kConfigFormatFix);
 
-      // Create and commit a pretty file without running lint_staged
+      // Create and commit a formatted file without running lint_staged
       // This way the file will be available for the next step
-      await project.writeFile('lib/main.dart', kFormattedDart);
-      await project.execGit(['add', '.']);
-      await project.execGit(['commit', '-m committed pretty file']);
+      await project.fs.write('lib/main.dart', kFormattedDart);
+      await project.git.run(['add', '.']);
+      await project.git.run(['commit', '-m committed formatted file']);
 
       // Edit file to be unformatted
-      await project.removeFile('lib/main.dart');
-      await project.writeFile('lib/main.dart', kUnFormattedDart);
-      await project.execGit(['add', 'lib/main.dart']);
+      await project.fs.remove('lib/main.dart');
+      await project.fs.write('lib/main.dart', kUnFormattedDart);
+      await project.git.run(['add', 'lib/main.dart']);
 
       // Run lint_staged to automatically format the file
       // Here we also pass '--allow-empty' to gitCommit because this part is not the full lint_staged
@@ -66,12 +65,10 @@ void main() {
           completes);
 
       // Nothing was wrong so the empty commit is created
-      expect(
-          await project.execGit(['rev-list', '--count', 'HEAD']), equals('3'));
-      expect(await project.execGit(['log', '-1', '--pretty=%B']),
-          contains('test'));
-      expect(await project.execGit(['diff', '-1']), equals(''));
-      expect(await project.readFile('lib/main.dart'), equals(kFormattedDart));
+      expect(await project.git.commitCount, equals(3));
+      expect(await project.git.lastCommit, contains('test'));
+      expect(await project.git.diff(['-1']), equals(''));
+      expect(await project.fs.read('lib/main.dart'), equals(kFormattedDart));
     });
   });
 }
