@@ -9,19 +9,17 @@ void main() {
   group('lint_staged', () {
     test('supports overriding file list using --diff', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
       // Commit unformatted file
-      await project.writeFile('lib/main.dart', kUnFormattedDart);
-      await project.execGit(['add', '.']);
-      await project.execGit(['commit', '-m unformatted']);
+      await project.fs.write('lib/main.dart', kUnFormattedDart);
+      await project.git.run(['add', '.']);
+      await project.git.run(['commit', '-m unformatted']);
 
-      final hashes = (await project.execGit(['log', '--format=format:%H']))
-          .trim()
-          .split('\n');
+      final hashes = await project.git.hashes;
       expect(hashes.length, 2);
 
       // Run lint_staged with `--diff` between the two commits.
@@ -29,7 +27,7 @@ void main() {
       final passed = await lintStaged(
         diff: ['${hashes[1]}..${hashes[0]}'],
         stash: false,
-        workingDirectory: project.dir,
+        workingDirectory: project.path,
       );
       // lint_staged failed because commit diff contains unformatted file
       expect(passed, isFalse);
@@ -37,20 +35,20 @@ void main() {
 
     test('supports overriding default --diff-filter', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
       // Stage unformatted file
-      await project.writeFile('lib/main.dart', kUnFormattedDart);
-      await project.execGit(['add', '.']);
+      await project.fs.write('lib/main.dart', kUnFormattedDart);
+      await project.git.run(['add', '.']);
 
       // Run lint_staged with `--diff-filter=D` to include only deleted files.
       final passed = await lintStaged(
         diffFilter: 'D',
         stash: false,
-        workingDirectory: project.dir,
+        workingDirectory: project.path,
       );
       // lint_staged passed because no matching (deleted) files
       expect(passed, isTrue);
