@@ -9,42 +9,42 @@ void main() {
     test('does not resurrect removed files due to git bug when tasks pass',
         () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
-      await project.removeFile('README.md'); // Remove file from previous commit
-      await project.writeFile('lib/main.dart', kFormattedDart);
-      await project.execGit(['add', 'lib/main.dart']);
+      await project.fs.remove('README.md'); // Remove file from previous commit
+      await project.fs.write('lib/main.dart', kFormattedDart);
+      await project.git.run(['add', 'lib/main.dart']);
 
       await project.gitCommit();
 
-      expect(await project.existsFile('README.md'), isFalse);
+      expect(await project.fs.exists('README.md'), isFalse);
     });
 
     test('does not resurrect removed files in complex case', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
       // Add file to index, and remove it from disk
-      await project.writeFile('lib/main.dart', kFormattedDart);
-      await project.execGit(['add', 'lib/main.dart']);
-      await project.removeFile('lib/main.dart');
+      await project.fs.write('lib/main.dart', kFormattedDart);
+      await project.git.run(['add', 'lib/main.dart']);
+      await project.fs.remove('lib/main.dart');
 
       // Rename file in index, and remove it from disk
-      final readme = await project.readFile('README.md');
-      await project.removeFile('README.md');
-      await project.execGit(['add', 'README.md']);
-      await project.writeFile('README_NEW.md', readme!);
-      await project.execGit(['add', 'README_NEW.md']);
-      await project.removeFile('README_NEW.md');
+      final readme = await project.fs.read('README.md');
+      await project.fs.remove('README.md');
+      await project.git.run(['add', 'README.md']);
+      await project.fs.write('README_NEW.md', readme!);
+      await project.git.run(['add', 'README_NEW.md']);
+      await project.fs.remove('README_NEW.md');
 
       expect(
-          await project.execGit(['status', '--porcelain']),
+          await project.git.status(['--porcelain']),
           contains('RD README.md -> README_NEW.md\n'
               'AD lib/main.dart\n'
               '?? pubspec.yaml'));
@@ -52,29 +52,29 @@ void main() {
       await project.gitCommit();
 
       expect(
-          await project.execGit(['status', '--porcelain']),
+          await project.git.status(['--porcelain']),
           contains(' D README_NEW.md\n'
               ' D lib/main.dart\n'
               '?? pubspec.yaml'));
 
-      expect(await project.existsFile('lib/main.dart'), isFalse);
-      expect(await project.existsFile('README_NEW.md'), isFalse);
+      expect(await project.fs.exists('lib/main.dart'), isFalse);
+      expect(await project.fs.exists('README_NEW.md'), isFalse);
     });
 
     test('does not resurrect removed files due to git bug when tasks fail',
         () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
-      await project.removeFile('README.md'); // Remove file from previous commit
-      await project.writeFile('lib/main.dart', kUnFormattedDart);
-      await project.execGit(['add', 'lib/main.dart']);
+      await project.fs.remove('README.md'); // Remove file from previous commit
+      await project.fs.write('lib/main.dart', kUnFormattedDart);
+      await project.git.run(['add', 'lib/main.dart']);
 
       expect(
-          await project.execGit(['status', '--porcelain']),
+          await project.git.status(['--porcelain']),
           contains(' D README.md\n'
               'A  lib/main.dart\n'
               '?? pubspec.yaml'));
@@ -82,12 +82,12 @@ void main() {
       await expectLater(project.gitCommit(allowEmpty: true), throwsException);
 
       expect(
-          await project.execGit(['status', '--porcelain']),
+          await project.git.status(['--porcelain']),
           contains(' D README.md\n'
               'A  lib/main.dart\n'
               '?? pubspec.yaml'));
 
-      expect(await project.existsFile('README.md'), isFalse);
+      expect(await project.fs.exists('README.md'), isFalse);
     });
   });
 }
