@@ -8,34 +8,32 @@ void main() {
   group('lint_staged', () {
     test('works when amending previous commit with unstaged changes', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
       // Edit file from previous commit
-      await project.appendFile('README.md', '\n## Amended\n');
-      await project.execGit(['add', 'README.md']);
+      await project.fs.append('README.md', '\n## Amended\n');
+      await project.git.run(['add', 'README.md']);
 
       // Edit again, but keep it unstaged
-      await project.appendFile('README.md', '\n## Edited\n');
-      await project.appendFile('lib/main.dart', kFormattedDart);
+      await project.fs.append('README.md', '\n## Edited\n');
+      await project.fs.append('lib/main.dart', kFormattedDart);
 
       // Run lint_staged with `dart format --set-exit-if-changed` and commit formatted file
       await project.gitCommit(gitCommitArgs: ['--amend', '--no-edit']);
 
       // Nothing is wrong, so the commit was amended
+      expect(await project.git.commitCount, equals(1));
+      expect(await project.git.lastCommit, contains('initial commit'));
       expect(
-          await project.execGit(['rev-list', '--count', 'HEAD']), equals('1'));
-      expect(await project.execGit(['log', '-1', '--pretty=%B']),
-          contains('initial commit'));
-      expect(
-          await project.readFile('README.md'),
+          await project.fs.read('README.md'),
           contains('# Test\n'
               '\n## Amended\n'
               '\n## Edited\n'));
-      expect(await project.readFile('lib/main.dart'), equals(kFormattedDart));
-      final status = await project.execGit(['status']);
+      expect(await project.fs.read('lib/main.dart'), equals(kFormattedDart));
+      final status = await project.git.status();
       expect(status, contains('modified:   README.md'));
       expect(status, contains('lib/'));
       expect(status, contains('no changes added to commit'));
