@@ -7,50 +7,48 @@ void main() {
   group('lint_staged', () {
     test('handles merge conflicts', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      final defaultBranch =
-          await project.execGit(['rev-parse', '--abbrev-ref', 'HEAD']);
+      final branch = project.git.currentBranch;
 
       final fileInBranchA = 'String foo = "foo";\n';
       final fileInBranchB = 'String foo="bar";\n';
       final fileInBranchBFixed = 'String foo = "bar";\n';
 
       // Create one branch
-      await project.execGit(['checkout', '-b', 'branch-a']);
-      await project.appendFile('lib/main.dart', fileInBranchA);
-      await project.appendFile('pubspec.yaml', kConfigFormatFix);
-      await project.execGit(['add', '.']);
+      await project.git.run(['checkout', '-b', 'branch-a']);
+      await project.fs.append('lib/main.dart', fileInBranchA);
+      await project.fs.append('pubspec.yaml', kConfigFormatFix);
+      await project.git.run(['add', '.']);
 
       await project.gitCommit(gitCommitArgs: ['-m commit a']);
 
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchA));
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchA));
 
-      await project.execGit(['checkout', defaultBranch]);
+      await project.git.run(['checkout', branch]);
 
       // Create another branch
-      await project.execGit(['checkout', '-b', 'branch-b']);
-      await project.appendFile('lib/main.dart', fileInBranchB);
-      await project.appendFile('pubspec.yaml', kConfigFormatFix);
-      await project.execGit(['add', '.']);
+      await project.git.run(['checkout', '-b', 'branch-b']);
+      await project.fs.append('lib/main.dart', fileInBranchB);
+      await project.fs.append('pubspec.yaml', kConfigFormatFix);
+      await project.git.run(['add', '.']);
       await project.gitCommit(gitCommitArgs: ['-m commit b']);
       expect(
-          await project.readFile('lib/main.dart'), equals(fileInBranchBFixed));
+          await project.fs.read('lib/main.dart'), equals(fileInBranchBFixed));
 
       // Merge first branch
-      await project.execGit(['checkout', defaultBranch]);
-      await project.execGit(['merge', 'branch-a']);
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchA));
-      expect(await project.execGit(['log', '-1', '--pretty=%B']),
-          contains('commit a'));
+      await project.git.run(['checkout', branch]);
+      await project.git.run(['merge', 'branch-a']);
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchA));
+      expect(await project.git.lastCommit, contains('commit a'));
 
       // Merge second branch, causing merge conflict
-      final merge = project.execGit(['merge', 'branch-b']);
+      final merge = project.git.run(['merge', 'branch-b']);
       await expectLater(merge, throwsException);
 
       expect(
-          await project.readFile('lib/main.dart'),
+          await project.fs.read('lib/main.dart'),
           contains('<<<<<<< HEAD\n'
               'String foo = "foo";\n'
               '=======\n'
@@ -59,68 +57,65 @@ void main() {
               ''));
 
       // Fix conflict and commit using lint_staged
-      await project.writeFile('lib/main.dart', fileInBranchB);
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchB));
-      await project.execGit(['add', '.']);
+      await project.fs.write('lib/main.dart', fileInBranchB);
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchB));
+      await project.git.run(['add', '.']);
 
       await project.gitCommit(gitCommitArgs: ['--no-edit']);
 
       // Nothing is wrong, so a new commit is created and file is pretty
-      expect(
-          await project.execGit(['rev-list', '--count', 'HEAD']), equals('4'));
-      final log = await project.execGit(['log', '-1', '--pretty=%B']);
+      expect(await project.git.commitCount, equals(4));
+      final log = await project.git.lastCommit;
       expect(log, contains('Merge branch \'branch-b\''));
       expect(log, contains('Conflicts:'));
       expect(log, contains('lib/main.dart'));
       expect(
-          await project.readFile('lib/main.dart'), equals(fileInBranchBFixed));
+          await project.fs.read('lib/main.dart'), equals(fileInBranchBFixed));
     });
     test('handles merge conflict when task errors', () async {
       final project = IntegrationProject();
-      print('dir: ${project.dir}');
+      print('dir: ${project.path}');
       await project.setup();
 
-      final defaultBranch =
-          await project.execGit(['rev-parse', '--abbrev-ref', 'HEAD']);
+      final branch = project.git.currentBranch;
 
       final fileInBranchA = 'String foo = "foo";\n';
       final fileInBranchB = 'String foo="bar";\n';
       final fileInBranchBFixed = 'String foo = "bar";\n';
 
       // Create one branch
-      await project.execGit(['checkout', '-b', 'branch-a']);
-      await project.appendFile('lib/main.dart', fileInBranchA);
-      await project.appendFile('pubspec.yaml', kConfigFormatFix);
-      await project.execGit(['add', '.']);
+      await project.git.run(['checkout', '-b', 'branch-a']);
+      await project.fs.append('lib/main.dart', fileInBranchA);
+      await project.fs.append('pubspec.yaml', kConfigFormatFix);
+      await project.git.run(['add', '.']);
 
       await project.gitCommit(gitCommitArgs: ['-m commit a']);
 
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchA));
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchA));
 
-      await project.execGit(['checkout', defaultBranch]);
+      await project.git.run(['checkout', branch]);
 
       // Create another branch
-      await project.execGit(['checkout', '-b', 'branch-b']);
-      await project.appendFile('lib/main.dart', fileInBranchB);
-      await project.appendFile('pubspec.yaml', kConfigFormatFix);
-      await project.execGit(['add', '.']);
+      await project.git.run(['checkout', '-b', 'branch-b']);
+      await project.fs.append('lib/main.dart', fileInBranchB);
+      await project.fs.append('pubspec.yaml', kConfigFormatFix);
+      await project.git.run(['add', '.']);
       await project.gitCommit(gitCommitArgs: ['-m commit b']);
       expect(
-          await project.readFile('lib/main.dart'), equals(fileInBranchBFixed));
+          await project.fs.read('lib/main.dart'), equals(fileInBranchBFixed));
 
       // Merge first branch
-      await project.execGit(['checkout', defaultBranch]);
-      await project.execGit(['merge', 'branch-a']);
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchA));
-      expect(await project.execGit(['log', '-1', '--pretty=%B']),
-          contains('commit a'));
+      await project.git.run(['checkout', branch]);
+      await project.git.run(['merge', 'branch-a']);
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchA));
+      expect(await project.git.lastCommit, contains('commit a'));
 
       // Merge second branch, causing merge conflict
       await expectLater(
-          project.execGit(['merge', 'branch-b']), throwsException);
+          project.git.run(['merge', 'branch-b']), throwsException);
 
       expect(
-          await project.readFile('lib/main.dart'),
+          await project.fs.read('lib/main.dart'),
           contains('<<<<<<< HEAD\n'
               'String foo = "foo";\n'
               '=======\n'
@@ -129,20 +124,19 @@ void main() {
               ''));
 
       // Fix conflict and commit using lint_staged
-      await project.writeFile('lib/main.dart', fileInBranchB);
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchB));
-      await project.execGit(['add', '.']);
+      await project.fs.write('lib/main.dart', fileInBranchB);
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchB));
+      await project.git.run(['add', '.']);
 
-      await project.writeFile('pubspec.yaml', kConfigFormatExit);
+      await project.fs.write('pubspec.yaml', kConfigFormatExit);
 
       await expectLater(project.gitCommit(), throwsException);
 
       // Something went wrong, so lintStaged failed and merge is still going
-      expect(
-          await project.execGit(['rev-list', '--count', 'HEAD']), equals('2'));
-      expect(await project.execGit(['status']),
+      expect(await project.git.commitCount, equals(2));
+      expect(await project.git.status(),
           contains('All conflicts fixed but you are still merging'));
-      expect(await project.readFile('lib/main.dart'), equals(fileInBranchB));
+      expect(await project.fs.read('lib/main.dart'), equals(fileInBranchB));
     });
   });
 }
